[RPC] IOS RPC (#261)

+++ src/runtime/module_util.cc
@@ -19,7 +19,7 @@ void ImportModuleBlob(const char* mblob, std::vector<Module>* mlist) {
     nbytes |=  (c & 0xffUL) << (i * 8);
   }
   dmlc::MemoryFixedSizeStream fs(
-      const_cast<char*>(mblob + sizeof(nbytes)), nbytes);
+      const_cast<char*>(mblob + sizeof(nbytes)), static_cast<size_t>(nbytes));
   dmlc::Stream* stream = &fs;
   uint64_t size;
   CHECK(stream->Read(&size));
